Extract common message bus behaviour tests
